refactor: move streamTabInfo and worktreeInfo out of agent/index
`streamTabInfo.ts` and `worktreeInfo.ts` had no business being in the
agent registry module (`src/agent/index/`). The registry's job is
agent discovery and lookup; these two files serve a completely different
purpose:

- `worktreeInfo.ts` is a pure git utility (resolves branch name + dirty
  status for any working directory). It now lives in `src/utils/git/`
  alongside other system utilities and carries no agent dependency.

- `streamTabInfo.ts` builds the `StreamTabInfo` UI object consumed by
  the progress view. It now lives in
  `src/shared/progressView/backend/` next to its primary consumer
  (`streamInfoUtils.ts`), cutting the cross-module hop to a relative
  import.

`streamInfoUtils.ts` is updated to import both from their canonical
locations. The `@agent/index` barrel re-exports `buildStreamTabInfo`
for backward compatibility while callers migrate.
