+++ src/shared/progressView/backend/streamInfoUtils.ts
@@ -1,8 +1,8 @@
-import { buildStreamTabInfo } from '@agent/index/streamTabInfo';
+import { buildStreamTabInfo } from './streamTabInfo';
 import {
   peekWorktreeInfo,
   resolveWorktreeInfo,
-} from '@agent/index/worktreeInfo';
+} from '@utils/git/worktreeInfo';
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import type { AgentCategoryFilter, StreamTabInfo } from '@shared/schemas';
 import { filterNotNull } from '@utils/core';
