+++ src/utils/git/worktreeInfo.ts
